+++ types/node_set.go
@@ -5,18 +5,18 @@ import (
 )
 
 type NodeSet struct {
-	GroupName string
-	Name      string
-	Mode      string
+	GroupName string `cty:"group_name"`
+	Name      string `cty:"name"`
+	Mode      string `cty:"mode"`
 	// Index is a node set counter over all created node sets.
-	Index int
+	Index int `cty:"index"`
 	// RelativeIndex is a counter relative to current node set group. Related to GroupName.
 	RelativeIndex int
 	// GroupIndex is a index of the group where this node set belongs to. Related to GroupName.
 	GroupIndex      int
-	Vega            VegaNode
-	Tendermint      TendermintNode
-	DataNode        *DataNode
+	Vega            VegaNode       `cty:"vega"`
+	Tendermint      TendermintNode `cty:"tendermint"`
+	DataNode        *DataNode      `cty:"data_node"`
 	Visor           *Visor
 	NomadJobRaw     *string `json:",omitempty"`
 	PreGenerateJobs []NomadJob