+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - [279](https://github.com/vegaprotocol/vegacapsule/issues/279) Use Vega binary only instead of `vegawallet` and `data-node` seperatly. Allow to define a specific vega binary version to node set and wallet
 - [270](https://github.com/vegaprotocol/vegacapsule/pull/270) Add vega isolated wallet id and the tendermint info into the `vegacapsule nodes ls` command.
 - [263](https://github.com/vegaprotocol/vegacapsule/issues/263) Add pre start probes for node sets
+- [288](https://github.com/vegaprotocol/vegacapsule/issues/288) Support Docker volumes mounts and add HCL configs templating
 
 ### 🐛 Fixes
 - [167](https://github.com/vegaprotocol/vegacapsule/issues/167) Fix validators filter in tendermint generator